server/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autoflake==2.2.1
billiard==4.1.0
boto3==1.28.32
botocore==1.31.32
celery==5.3.1
certifi==2015.4.28
cffi==1.15.1
chardet==3.0.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
convertdate==2.4.0
cryptography==41.0.3
dataclass-type-validator==0.1.2
django-cors-headers==4.3.1
dataclasses==0.6
distlib==0.3.7
Django==4.2.4
django-crispy-forms==1.10.0
django-datatables-too==0.1.4
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-filter==23.2
django-pgviews-redux==0.9.4
django-sendgrid-v5==1.2.2
django-storages==1.13.2
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.4
Faker==19.3.0
filelock==3.12.2
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
holidays==0.31
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
inflection==0.5.1
install==1.3.5
jmespath==1.0.1
JSON-log-formatter==0.5.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kombu==5.3.1
logging-utilities==4.0.0
lunardate==0.2.0
minio==7.1.16
Pillow==10.0.0
pipenv==2023.7.23
platformdirs==3.10.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
pyluach==2.2.0
PyMeeus==0.5.12
python-dateutil==2.8.2
python-dotenv==1.0.0
python-http-client==3.3.7
pytz==2023.3
PyYAML==6.0.1
querystring-parser==1.2.4
redis==5.0.0
referencing==0.30.2
rfc3986==1.5.0
rpds-py==0.9.2
s3transfer==0.6.2
sendgrid==6.10.0
sentry-sdk==1.40.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
tblib==2.0.0
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
virtualenv==20.24.2
virtualenv-clone==0.5.7
wcwidth==0.2.6
workalendar==17.0.0
